chore: rename src directory to basilisk directory for a top level python package
