+++ basilisk/conversation.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from enum import Enum
 from typing import Literal
 from pydantic import BaseModel, Field
-from provideraimodel import ProviderAIModel
+from .provideraimodel import ProviderAIModel
 
 
 class MessageRoleEnum(Enum):
