+++ basilisk/globalvars.py
