+++ basilisk/gui/aboutdialog.py
