+++ basilisk/imagehelper.py
