+++ basilisk/providerengine/__init__.py
@@ -1,6 +1,3 @@
-import sys
-
-sys.path.append("..")
 from .baseengine import BaseEngine as BaseEngine
 from .mistralaiengine import MistralAIEngine as MistralAIEngine
 from .openaiengine import OpenAIEngine as OpenAIEngine
