+++ basilisk/res/locale/fr/LC_MESSAGES/basiliskLLM.po
